change datatype for combined_img variable in merge_image_browser function

Former-commit-id: 83889d38fe5faed207a866ed8697923a12ad532a

+++ code/10-image_stitching/01-samui_test.py
@@ -195,7 +195,7 @@ def merge_image_browser(sample_info, theta, trans_img, max0, max1):
     #   maximal coordinates in each dimension of each rotated and translated
     #   image
     combined_img = np.full(
-        (max0 + 1, max1 + 1, 3), 240, dtype = np.uint8
+        (max0 + 1, max1 + 1, 3), 240, dtype = np.float16
     )
     weights = np.zeros((max0 + 1, max1 + 1, 1), dtype = np.float64)
